IMPORTANT TILESHEETS CHANGE
Hello GitHub Gather users!
Today we're making a big change to the way we organize tilesheets in our folders. The result of this will break your existing Tiled files, or any files that look for the file using the old file path. We are doing this to make a clear dilineation of old floor/wall style and new floor/wall style as we make way for the new Mapmaker system coming to Gather. The tilesheets that are now contemporary can be found in the "Floors" folder within the "Tilesets" folder. I recommend trying those out in your future maps, as we are proud and they are more colorful and subtle than our old floors. In order to fix your maps that rely on old tilesheets, open Tiled, acknowledge any error message by relocating the tilesheet in the folder  called "deprecated tilesheets".